requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
